refactor: split OpenAI module into directory
