+++ src/lib/OpenAI/client.ts
@@ -1,0 +1,5 @@
+import OpenAI from "openai";
+
+export default function createClient(apiKey: string): OpenAI {
+  return new OpenAI({ apiKey, dangerouslyAllowBrowser: true });
+}
